refactor: :recycle: lottie-web 사용 코드 Lottie 컴포넌트로 대체
