+++ src/assets/lottie/index.ts
@@ -1,0 +1,4 @@
+export { default as loadingLottie } from './loading.json';
+export { default as loginLottie } from './login.json';
+export { default as signupLottie } from './signup.json';
+export { default as splashLottie } from './splash.json';
