+++ src/pages/home/index.page.tsx
@@ -1,35 +1,15 @@
-import { useEffect, useRef } from 'react';
-import Lottie from 'lottie-web';
-
-import splashLottie from '@src/assets/lotti/splash.json';
 import GuestGuard from '@src/features/auth/components/GuestGuard';
 import { NextPageWithLayout } from '@src/shared/types/next';
+import { splashLottie } from '@src/assets/lottie';
 
 import * as S from './home.styles';
 
 const Home: NextPageWithLayout = () => {
-  const lottieRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (lottieRef === null) {
-      return;
-    }
-
-    Lottie.loadAnimation({
-      container: lottieRef.current,
-      renderer: 'svg',
-      loop: false,
-      autoplay: true,
-      animationData: splashLottie,
-      path: '',
-    });
-  }, []);
-
   return (
     <S.Wrapper>
       <S.Title>내 메모를 채팅처럼</S.Title>
       <img src="/images/logo.png" alt="메모쳇" width={222} height={30} />
-      <S.Lottie ref={lottieRef} />
+      <S.Lottie animationData={splashLottie} autoplay loop />
       <S.LoginBtn href="/signin">로그인</S.LoginBtn>
       <S.Box>
         <S.TextLink href="/signup" passHref>
