+++ src/pages/signin/complete/index.page.tsx
@@ -1,31 +1,14 @@
-import { useEffect, useRef } from 'react';
-import Lottie from 'lottie-web';
 import { useRouter } from 'next/router';
 import { NextPage } from 'next';
 
-import loginLottie from '@src/assets/lotti/login.json';
 import { Button } from '@src/shared/components';
+import { loginLottie } from '@src/assets/lottie';
+import Lottie from '@src/shared/components/Lottie';
 
 import * as S from './complete.styles';
 
 const SignInComplete: NextPage = () => {
-  const lottieRef = useRef<HTMLDivElement>(null);
   const router = useRouter();
-
-  useEffect(() => {
-    if (lottieRef === null) {
-      return;
-    }
-
-    Lottie.loadAnimation({
-      container: lottieRef.current as HTMLDivElement,
-      renderer: 'svg',
-      loop: false,
-      autoplay: true,
-      animationData: loginLottie,
-      path: '',
-    });
-  }, []);
 
   const handleOnStartBtnClick = () => {
     router.push('/rooms');
@@ -36,7 +19,7 @@ const SignInComplete: NextPage = () => {
       <S.Header>
         <S.Title>로그인을 완료하였습니다!</S.Title>
       </S.Header>
-      <S.Content ref={lottieRef} />
+      <Lottie animationData={loginLottie} autoplay />
       <S.ButtonContainer>
         <Button onClick={handleOnStartBtnClick}>시작하기</Button>
       </S.ButtonContainer>
