+++ src/shared/components/Loading/Loading.tsx
@@ -1,32 +1,14 @@
-import Lottie from 'lottie-web';
 import { useRouter } from 'next/router';
 import nProgress from 'nprogress';
-import { useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 import { useEffectOnce } from 'react-use';
 
-import loadingLottie from '@src/assets/lotti/loading.json';
-
-import * as S from './Loading.styles';
 import { LoadingProps } from './Loading.types';
+import * as S from './Loading.styles';
 
 const Loading = ({ initialLoadingState = false }: LoadingProps) => {
   const router = useRouter();
-  const lottieRef = useRef<HTMLDivElement>(null);
   const [isLoading, setIsLoading] = useState(initialLoadingState);
-
-  useEffect(() => {
-    if (lottieRef === null || !isLoading) {
-      return;
-    }
-    Lottie.loadAnimation({
-      container: lottieRef.current as HTMLDivElement,
-      renderer: 'svg',
-      loop: true,
-      autoplay: true,
-      animationData: loadingLottie,
-      path: '',
-    });
-  }, [isLoading]);
 
   useEffectOnce(() => {
     if (!router.isReady) {
@@ -58,7 +40,7 @@ const Loading = ({ initialLoadingState = false }: LoadingProps) => {
   return (
     <S.Wrapper>
       {/* <S.Dim open={isLoading} /> */}
-      {/* <S.Lottie ref={lottieRef} /> */}
+      {/* <S.Lottie animationData={loadingLottie} /> */}
     </S.Wrapper>
   );
 };
